various cleanup
- combine natural event util files
- dont filter by proj/extent ui side
